removal/requirements.txt
oyaml==0.9
comet-ml==3.0.2
pillow==6.2.2
numpy==1.18.1
torch==1.1.0
torchvision==0.3.0
